# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Southern Star Diamonds Ring Builder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
  code?: string;
}

export interface SettingVariant {
  sku: string;
  price: number;
  metal: string;
  shape: string;
  image: string;
  /** @nullable */
  videoUrl?: string | null;
}

export interface Setting {
  handle: string;
  title: string;
  description: string;
  productType: string;
  tags: string[];
  images: string[];
  variants: SettingVariant[];
  /** Classification fields merged from CSV */
  settingHeight?: string;
  bandType?: string;
  settingType?: string;
}

export interface SettingListResponse {
  items: Setting[];
  total: number;
  page: number;
  pageSize: number;
  pages: number;
}

export interface DiamondCertificate {
  /** @nullable */
  id?: string | null;
  /** @nullable */
  lab?: string | null;
  shape: string;
  /** @nullable */
  certNumber?: string | null;
  /** @nullable */
  cut?: string | null;
  carats: number;
  clarity: string;
  /** @nullable */
  polish?: string | null;
  /** @nullable */
  symmetry?: string | null;
  color: string;
  /** @nullable */
  length?: number | null;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  depth?: number | null;
  /** @nullable */
  table?: number | null;
  /** @nullable */
  depthPercentage?: number | null;
}

export interface Diamond {
  id: string;
  price: number;
  /** @nullable */
  discount?: number | null;
  /** @nullable */
  supplierPrice?: number | null;
  /** @nullable */
  video?: string | null;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  supplierStockId?: string | null;
  certificate: DiamondCertificate;
}

export type DiamondSearchInputSortDir = typeof DiamondSearchInputSortDir[keyof typeof DiamondSearchInputSortDir];


export const DiamondSearchInputSortDir = {
  asc: 'asc',
  desc: 'desc',
} as const;

export interface DiamondSearchInput {
  shapes?: string[];
  caratMin?: number;
  caratMax?: number;
  priceMin?: number;
  priceMax?: number;
  colors?: string[];
  clarities?: string[];
  cuts?: string[];
  labGrown?: boolean;
  offset?: number;
  limit?: number;
  sortBy?: string;
  sortDir?: DiamondSearchInputSortDir;
}

export interface DiamondSearchResponse {
  items: Diamond[];
  totalCount: number;
  offset: number;
  limit: number;
}

export interface DiamondValidateInput {
  nivodaId: string;
}

export interface DiamondValidateResponse {
  available: boolean;
  /** @nullable */
  currentPrice?: number | null;
  priceChanged: boolean;
  diamond?: Diamond;
}

export interface CartLineItem {
  settingHandle: string;
  settingTitle: string;
  settingSku: string;
  metal: string;
  shape: string;
  ringSize: string;
  /** @nullable */
  engraving?: string | null;
  settingPrice: number;
  nivodaId: string;
  diamondCarat: number;
  diamondColor: string;
  diamondClarity: string;
  /** @nullable */
  diamondCut?: string | null;
  /** @nullable */
  diamondCert?: string | null;
  /** @nullable */
  diamondLab?: string | null;
  diamondPrice: number;
}

export interface CartInput {
  lineItem: CartLineItem;
}

export interface CartResponse {
  checkoutUrl: string;
  /** @nullable */
  cartId?: string | null;
}

export interface MetalOption {
  label: string;
  karat: string;
  tone: string;
}

export interface ShapeOption {
  label: string;
  nivodaKey: string;
}

export type PublicConfigDiamondTypes = {
  labGrown: boolean;
  natural: boolean;
};

export interface PublicConfig {
  metals: MetalOption[];
  shapes: ShapeOption[];
  diamondTypes: PublicConfigDiamondTypes;
  currency: string;
  engravingEnabled: boolean;
  ringSizes: string[];
}

export type ListSettingsParams = {
style?: string;
metal?: string;
shape?: string;
minPrice?: number;
maxPrice?: number;
page?: number;
pageSize?: number;
sort?: ListSettingsSort;
/**
 * Search query
 */
q?: string;
};

export type ListSettingsSort = typeof ListSettingsSort[keyof typeof ListSettingsSort];


export const ListSettingsSort = {
  'price-asc': 'price-asc',
  'price-desc': 'price-desc',
  'name-asc': 'name-asc',
} as const;